+++ src/main/java/com/honeyboard/api/auth/service/VerificationServiceImpl.java
@@ -1,4 +1,4 @@
-package com.honeyboard.api.auth.model.service;
+package com.honeyboard.api.auth.service;
 
 import java.util.Random;
 import java.util.concurrent.TimeUnit;
