+++ src/main/java/com/honeyboard/api/auth/util/OAuth2UserInfoFactory.java
@@ -1,4 +1,4 @@
-package com.honeyboard.api.auth.model.util;
+package com.honeyboard.api.auth.util;
 
 import java.util.Map;
