+++ src/main/java/com/honeyboard/api/project/finale/controller/FinaleTeamController.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 import com.honeyboard.api.project.finale.model.FinaleProjectResponse;
 import com.honeyboard.api.project.finale.service.FinaleTeamService;
-import com.honeyboard.api.user.model.mapper.UserMapper;
-import com.honeyboard.api.user.model.service.UserService;
+import com.honeyboard.api.user.mapper.UserMapper;
+import com.honeyboard.api.user.service.UserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
