+++ src/main/java/com/honeyboard/api/user/service/UserService.java
@@ -1,4 +1,4 @@
-package com.honeyboard.api.user.model.service;
+package com.honeyboard.api.user.service;
 
 import com.honeyboard.api.user.model.User;
